requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.9.1
asyncio==3.4.3
attrs==24.2.0
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.1
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.6
propcache==0.2.1
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
